refactor: 課程總價計算檔案搬遷
