+++ controllers/cartController/cartOperations.js
@@ -39,21 +39,6 @@ const getCourseData = async (courseIds, attributes) => {
   }
 }
 
-const calculateTotalPrice = (courseData) => {
-  let totalPrice = 0
-
-  courseData.forEach((course) => {
-    if (course.price < course.originPrice) {
-      totalPrice += course.price
-    } else {
-      totalPrice += course.originPrice
-    }
-  })
-
-  return totalPrice
-}
-
 module.exports = {
-  getCourseData,
-  calculateTotalPrice
+  getCourseData
 }
