+++ src/utils/calculate.js
@@ -1,0 +1,17 @@
+const calculateTotalPrice = (courseData) => {
+  let totalPrice = 0
+
+  courseData.forEach((course) => {
+    if (course.price < course.originPrice) {
+      totalPrice += course.price
+    } else {
+      totalPrice += course.originPrice
+    }
+  })
+
+  return totalPrice
+}
+
+module.exports = {
+  calculateTotalPrice
+}
